Improve logging/printed output
Add in a verbose mode that uses logging to output more
information when available. Also adjusts the output to
show which checks had which errors (which is useful to
know and is useful in general) as well as show the total
number of errors accumulated (and the specific break
downs).

Change-Id: Ia965aa7c7c3576230668217aa09cf61dc4ecd303