simplify `pyproject.toml` argument normalization (#75)
Instead of using the implementation of `from_ini()` as inspiration,
simply loop over the items in the `[tool.doc8]` dictionary and replace
hyphens in names with underscores.

There is a special case for the `ignore-path-errors` argument because
the elements of that list each need to be parsed.

I removed the extra handling for the `extensions` argument which
stripped whitespace from each element in the list.  On reflection, I
realized that the stripping was necessary for INI files due to the way
the `ConfigParser` works and not because it's necessary to guard
against extensions with whitespace.